Start to modularize our growing collection of tests.
